Extract remote compaction request attempts
