+++ codex-rs/core/src/compact_remote_v2_attempt.rs
@@ -1,0 +1,130 @@
+use std::sync::Arc;
+
+use super::RemoteCompactionV2Output;
+use super::run_remote_compaction_request_v2;
+use crate::Prompt;
+use crate::client::ModelClientSession;
+use crate::compact::CompactionAnalyticsDetails;
+use crate::compact_remote::trim_function_call_history_to_fit_context_window;
+use crate::responses_metadata::CodexResponsesRequestKind;
+use crate::responses_metadata::CompactionTurnMetadata;
+use crate::session::session::Session;
+use crate::session::step_context::StepContext;
+use crate::session::turn::built_tools;
+use crate::session::turn_context::TurnContext;
+use codex_protocol::error::Result as CodexResult;
+use codex_protocol::models::ResponseItem;
+use codex_protocol::protocol::TokenUsage;
+use codex_rollout_trace::CompactionTraceContext;
+use tokio_util::sync::CancellationToken;
+use tracing::info;
+
+pub(super) struct RemoteCompactV2Attempt {
+    pub(super) turn_context: Arc<TurnContext>,
+    pub(super) trace_input_history: Vec<ResponseItem>,
+    pub(super) prompt_input: Vec<ResponseItem>,
+    pub(super) compaction_output: ResponseItem,
+    pub(super) token_usage: Option<TokenUsage>,
+    pub(super) compaction_trace: CompactionTraceContext,
+}
+
+pub(super) async fn run_remote_compact_v2_attempt(
+    sess: &Arc<Session>,
+    step_context: &Arc<StepContext>,
+    client_session: &mut ModelClientSession,
+    compaction_item_id: &str,
+    compaction_metadata: CompactionTurnMetadata,
+    active_context_tokens_before: Option<i64>,
+    analytics_details: &mut CompactionAnalyticsDetails,
+) -> CodexResult<RemoteCompactV2Attempt> {
+    let turn_context = Arc::clone(&step_context.turn);
+    let compaction_trace = sess.services.rollout_thread_trace.compaction_trace_context(
+        turn_context.sub_id.as_str(),
+        compaction_item_id,
+        turn_context.model_info.slug.as_str(),
+        turn_context.provider.info().name.as_str(),
+    );
+    let mut history = sess.clone_history().await;
+    let base_instructions = sess.get_base_instructions().await;
+    let (rewritten_outputs, estimated_deleted_tokens) =
+        trim_function_call_history_to_fit_context_window(
+            &mut history,
+            turn_context.as_ref(),
+            &base_instructions,
+        );
+    if rewritten_outputs > 0 {
+        info!(
+            turn_id = %turn_context.sub_id,
+            rewritten_outputs,
+            "rewrote history outputs before remote compaction v2"
+        );
+    }
+    analytics_details.active_context_tokens_before = active_context_tokens_before;
+    if estimated_deleted_tokens > 0 {
+        let max_local_deleted_tokens = sess
+            .estimated_tokens_after_last_model_generated_item()
+            .await;
+        analytics_details.active_context_tokens_before =
+            active_context_tokens_before.map(|active_context_tokens_before| {
+                active_context_tokens_before
+                    .saturating_sub(estimated_deleted_tokens.min(max_local_deleted_tokens))
+            });
+    }
+
+    let trace_input_history = history.raw_items().to_vec();
+    let prompt_input = history.for_prompt(&turn_context.model_info.input_modalities);
+    let tool_router = built_tools(
+        sess.as_ref(),
+        step_context.as_ref(),
+        &CancellationToken::new(),
+    )
+    .await?;
+    let mut input = prompt_input.clone();
+    input.push(ResponseItem::CompactionTrigger {});
+    let prompt = Prompt {
+        input,
+        tools: tool_router.model_visible_specs(),
+        parallel_tool_calls: turn_context.model_info.supports_parallel_tool_calls,
+        base_instructions,
+        output_schema: None,
+        output_schema_strict: true,
+    };
+
+    let window_id = sess.current_window_id().await;
+    let responses_metadata = turn_context.turn_metadata_state.to_responses_metadata(
+        sess.installation_id.clone(),
+        window_id,
+        CodexResponsesRequestKind::Compaction(compaction_metadata),
+    );
+    let trace_attempt = compaction_trace.start_attempt(&serde_json::json!({
+        "model": turn_context.model_info.slug.as_str(),
+        "instructions": prompt.base_instructions.text.as_str(),
+        "input": &prompt.input,
+        "parallel_tool_calls": prompt.parallel_tool_calls,
+    }));
+    let compaction_output_result = run_remote_compaction_request_v2(
+        sess,
+        turn_context.as_ref(),
+        client_session,
+        &prompt,
+        &responses_metadata,
+    )
+    .await;
+    trace_attempt.record_result(
+        compaction_output_result
+            .as_ref()
+            .map(|output| std::slice::from_ref(&output.compaction_output)),
+    );
+    let RemoteCompactionV2Output {
+        compaction_output,
+        token_usage,
+    } = compaction_output_result?;
+    Ok(RemoteCompactV2Attempt {
+        turn_context,
+        trace_input_history,
+        prompt_input,
+        compaction_output,
+        token_usage,
+        compaction_trace,
+    })
+}
